+++ src/BusinessLogic/DraftShipment/Exceptions/DraftTaskMapExists.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Packlink\BusinessLogic\DraftShipment\Exceptions;
+
+use Logeecom\Infrastructure\Exceptions\BaseException;
+
+/**
+ * Class DraftTaskMapAlreadyExists.
+ *
+ * @package Packlink\BusinessLogic\DraftShipment\Exceptions
+ */
+class DraftTaskMapExists extends BaseException
+{
+}