+++ src/BusinessLogic/DraftShipment/Exceptions/DraftTaskMapNotFound.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Packlink\BusinessLogic\DraftShipment\Exceptions;
+
+use Logeecom\Infrastructure\Exceptions\BaseException;
+
+/**
+ * Class DraftTaskMapNotFound.
+ *
+ * @package Packlink\BusinessLogic\DraftShipment\Exceptions
+ */
+class DraftTaskMapNotFound extends BaseException
+{
+}